requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=1.40.0

# ── Scraping ──────────────────────────────────────────────────────────────────
playwright>=1.44.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# ── Google ────────────────────────────────────────────────────────────────────
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# ── Data ──────────────────────────────────────────────────────────────────────
pandas>=2.2.0
